refactor: migrate CLI modules from voyager-verifier to voyager crate
Strip voyager-verifier down to a pure API library by removing all CLI-only
modules (args, commands, config, wizard, output formatting, storage history,
notifications, error types, and the build script). Move those modules into the
voyager CLI crate along with a new file_collector, project_detection, license,
and errors layer. Add errors.rs and manifest.rs to the library surface.
Update Cargo.toml dependencies for both crates accordingly.
