+++ crates/voyager-verifier/Cargo.toml
@@ -12,7 +12,6 @@ version.workspace = true
 include = [
     "/src/**/*.rs",
     "/Cargo.toml",
-    "/build.rs",
 ]
 
 [package.metadata.cargo-package-metadata]
@@ -34,31 +33,20 @@ anyhow.workspace = true
 backon.workspace = true
 camino.workspace = true
 chrono.workspace = true
-clap.workspace = true
-colored.workspace = true
-dialoguer.workspace = true
-dirs.workspace = true
 itertools.workspace = true
 lazy_static.workspace = true
 log.workspace = true
-notify-rust = { workspace = true, optional = true }
 regex.workspace = true
 reqwest.workspace = true
-rusqlite.workspace = true
 scarb-metadata.workspace = true
 semver.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 serde_repr.workspace = true
-spdx.workspace = true
 thiserror.workspace = true
 toml.workspace = true
 url.workspace = true
 walkdir.workspace = true
-
-[features]
-default = ["notifications"]
-notifications = ["notify-rust"]
 
 [dev-dependencies]
 assert_fs.workspace = true
