+++ crates/voyager-verifier/src/core/mod.rs
@@ -1,3 +1,2 @@
 pub mod class_hash;
 pub mod project;
-pub mod verification;
