+++ crates/voyager-verifier/src/manifest.rs
