--- crates/voyager-verifier/src/utils/mod.rs
@@ -1,3 +1,0 @@
-pub mod errors;
-pub mod license;
-pub mod voyager;
