+++ crates/voyager-verifier/tests/error_message_tests.rs
@@ -4,8 +4,8 @@ use reqwest::StatusCode;
 use url::Url;
 use voyager_verifier::api::{ApiClientError, VerificationError};
 use voyager_verifier::core::class_hash::{ClassHash, ClassHashError};
+use voyager_verifier::errors::{MissingContract, RequestFailure};
 use voyager_verifier::filesystem::resolver;
-use voyager_verifier::utils::errors::{MissingContract, RequestFailure};
 
 #[test]
 fn test_missing_contract_error_with_suggestions() {
