+++ crates/voyager/src/license.rs
